+++ src/modules/pos.ts
@@ -25,6 +25,10 @@ type GetOrdersParams = AutoPaginatedParams<operations['pos_get_orders']['paramet
 
 type GetTaxesParams = AutoPaginatedParams<operations['pos_get_taxes']['parameters']['query']>;
 
+type GetModifiersParams = AutoPaginatedParams<
+    operations['pos_get_modifiers']['parameters']['query']
+>;
+
 const posFactory = {
     getLocations(options?: RawDataOption): RequestData<components['schemas']['POSLocationItem'][]> {
         return {
@@ -195,6 +199,17 @@ const posFactory = {
             params,
             method: 'get',
             url: `/consumers/{consumer_id}/pos/tax-rates`,
+            rawData: options?.rawData,
+        };
+    },
+    getModifiers(
+        params?: GetModifiersParams,
+        options?: RawDataOption
+    ): RequestData<components['schemas']['ModifiersItem'][]> {
+        return {
+            params,
+            method: 'get',
+            url: `/consumers/{consumer_id}/pos/modifiers`,
             rawData: options?.rawData,
         };
     },